+++ JavaCodes/Assignments/Arrays/ExampleofArray.java
@@ -1,0 +1,21 @@
+package vishwa;
+
+public class ExampleofArray {
+    public static void main(String[] args) {
+      /*
+        int arr[]={1,2,3,4,5,6,7,8,9}; //we assign directly.
+        System.out.print(arr[1]); // it prints 1st index.
+      */
+        /*
+        //for integer.
+        int arr[]; //declaration of arr.
+        arr=new int[5]; //Initialisation & we've to give arr size else it will show error.
+        System.out.println(arr[4]); // by default it is assigned with zeros [0 0 0 0 0]
+
+        //for string.
+        String arr[]=new String[5];
+        System.out.println(arr[1]); //prints null, default value.
+        */
+
+    }
+}
