+++ JavaCodes/Assignments/LoopsAndConditionals/src/vishwa/Discount.java
@@ -1,0 +1,7 @@
+package vishwa;
+import java.util.*;
+public class Discount {
+    public static void main(String[] args) {
+        
+    }
+}
